TagDetailService moved to new folder Tag 	modified:   app/Http/Controllers/api/Tag/TagDetailController.php 	renamed:    app/Service/TagDetailService.php -> app/Service/Tag/TagDetailService.php 	modified:   tests/Feature/Service/TagDetailServiceTest.php 	modified:   tests/Feature/TagDetailControllerTest.php
